Use new Routes API to hand directions to the LLM
The react-google-maps stuff doesn't give a convenient way to perform an API request in isolation. And it happens only when the Map is being shown in the chat, i.e. after the LLM finished responding. This is a bit inefficient, since now two requests are made, maybe something to optimize later.