Disable finding artists
When instructed to "play music by XYZ", the LLM most often searches the artist first, then uses the "play_top_tracks" functionality using the same string. Maybe I need to revert this when it turns out the search up front is useful when the transcription has garbled the artist name... but right now, it feels like a waste of latency.